// src/_proto/Protos/options.ts
export interface FieldStringLengthOptions {
    /**
     * @generated from protobuf field: google.protobuf.Int32Value minLength = 1;
     */
    minLength?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value maxLength = 2;
     */
    maxLength?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.StringValue errorMessage = 3;
     */
    errorMessage?: StringValue;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.FieldWordLengthOptions
 */
export interface FieldWordLengthOptions {
    /**
     * @generated from protobuf field: google.protobuf.Int32Value minWords = 1;
     */
    minWords?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value maxWords = 2;
     */
    maxWords?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.StringValue errorMessage = 3;
     */
    errorMessage?: StringValue;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.FieldRequiredOption
 */
export interface FieldRequiredOption {
    /**
     * @generated from protobuf field: bool value = 1;
     */
    value: boolean;
    /**
     * @generated from protobuf field: google.protobuf.StringValue errorMessage = 2;
     */
    errorMessage?: StringValue;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.FieldDisplayOption
 */
export interface FieldDisplayOption {
    /**
     * @generated from protobuf field: string name = 1;
     */
    name: string;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.FieldRangeOption
 */
export interface FieldRangeOption {
    /**
     * @generated from protobuf field: google.protobuf.Int32Value min = 1;
     */
    min?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value max = 2;
     */
    max?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.StringValue errorMessage = 3;
     */
    errorMessage?: StringValue;
}
// @generated message type with reflection information, may provide speed optimized methods
class FieldStringLengthOptions$Type extends MessageType<FieldStringLengthOptions> {
    constructor() {
        super("wuxiaworld.api.v2.FieldStringLengthOptions", [
            { no: 1, name: "minLength", kind: "message", T: () => Int32Value },
            { no: 2, name: "maxLength", kind: "message", T: () => Int32Value },
            { no: 3, name: "errorMessage", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.FieldStringLengthOptions
 */
export const FieldStringLengthOptions = new FieldStringLengthOptions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FieldWordLengthOptions$Type extends MessageType<FieldWordLengthOptions> {
    constructor() {
        super("wuxiaworld.api.v2.FieldWordLengthOptions", [
            { no: 1, name: "minWords", kind: "message", T: () => Int32Value },
            { no: 2, name: "maxWords", kind: "message", T: () => Int32Value },
            { no: 3, name: "errorMessage", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.FieldWordLengthOptions
 */
export const FieldWordLengthOptions = new FieldWordLengthOptions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FieldRequiredOption$Type extends MessageType<FieldRequiredOption> {
    constructor() {
        super("wuxiaworld.api.v2.FieldRequiredOption", [
            { no: 1, name: "value", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 2, name: "errorMessage", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.FieldRequiredOption
 */
export const FieldRequiredOption = new FieldRequiredOption$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FieldDisplayOption$Type extends MessageType<FieldDisplayOption> {
    constructor() {
        super("wuxiaworld.api.v2.FieldDisplayOption", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.FieldDisplayOption
 */
export const FieldDisplayOption = new FieldDisplayOption$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FieldRangeOption$Type extends MessageType<FieldRangeOption> {
    constructor() {
        super("wuxiaworld.api.v2.FieldRangeOption", [
            { no: 1, name: "min", kind: "message", T: () => Int32Value },
            { no: 2, name: "max", kind: "message", T: () => Int32Value },
            { no: 3, name: "errorMessage", kind: "message", T: () => StringValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.FieldRangeOption
 */
export const FieldRangeOption = new FieldRangeOption$Type();